Fix tests for 32bit machines

+++ test/broadcast.jl
@@ -966,7 +966,7 @@ p0 = copy(p)
     @test isequal(tuple.([1, 2], [3.0, missing]), [(1, 3.0), (2, missing)])
     @test Core.Compiler.return_type(broadcast, Tuple{typeof(tuple), Vector{Int},
                                                      Vector{Union{Float64, Missing}}}) ==
-        Union{Vector{Tuple{Int64, Missing}}, Vector{Tuple{Int64, Any}}, Vector{Tuple{Int64, Float64}}}
+        Union{Vector{Tuple{Int, Missing}}, Vector{Tuple{Int, Any}}, Vector{Tuple{Int, Float64}}}
     # Check that corner cases do not throw an error
     @test isequal(broadcast(x -> x === 1 ? nothing : x, [1, 2, missing]),
                   [nothing, 2, missing])
